捕获文件被其他线程删除的异常

+++ pytest_project/clear_download_dir.py
@@ -8,6 +8,7 @@ from pytest_project.config.conf import cm
 class Clear(object):
     download_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'download-dir')
     log_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'logs')
+    log = Log().get_log()
 
     def clear_download_files(self):
         for i in os.listdir(self.download_dir):
@@ -15,15 +16,16 @@ class Clear(object):
                 continue
             else:
                 try:
-                    os.remove(self.download_dir + os.sep + i)
+                    if os.path.exists((self.download_dir + os.sep + i)):
+                        os.remove(self.download_dir + os.sep + i)
                 except PermissionError:
                     time.sleep(5)
                     try:
                         os.remove(self.download_dir + os.sep + i)
-                    except PermissionError:
-                        Log().get_log().error('清空下载文件夹失败，手动清空')
                     except FileNotFoundError:
-                        Log().get_log().error('文件已被其他线程清除')
+                        self.log.error('文件已被其他线程清除')
+                    except PermissionError:
+                        self.log.error('清空下载文件夹失败，手动清空')
 
     def clear_log_file(self):
         for i in os.listdir(self.log_dir):
